* Updates to do padding by batch. * Refactor to move file management code into utilties so can be re-used
