+++ constants.py
@@ -1,5 +1,6 @@
 DIMENSIONS = 64
 DATASET = "train"
+DATASET_PERCENTAGE = 100
 VOCAB_SIZE = 16000
 NUM_OF_EPOCHS = 100
 LEARNING_RATE = 0.005
@@ -9,8 +10,9 @@ WANDB_ON = False
 NUM_HEADS=4
 NUM_LAYERS=1
 D_FF =  2048
-MAX_SEQ_LENGTH = 100
+MAX_SEQ_LENGTH = 1200
 DROPOUT = 0.1
 
 INPUT_SENTENCE_SIZE = 100000
+
