+++ train.py
@@ -8,8 +8,6 @@ import sentencepiece as spm
 import tqdm
 import constants
 import wandb
-import os
-import re
 import utilities
 
 # start a new wandb run to track this script
@@ -36,8 +34,8 @@ if constants.WANDB_ON:
 device = utilities.getDevice()
 print(f"Device = {device}")
 
-ds = dataset.TinyStoriesData("roneneldan/TinyStories", "train[:1%]", constants.MAX_SEQ_LENGTH)
-dl = torch.utils.data.DataLoader(ds, batch_size=constants.BATCH_SIZE, shuffle=True)
+ds = dataset.TinyStoriesData("roneneldan/TinyStories", f"train[:{constants.DATASET_PERCENTAGE}%]", constants.MAX_SEQ_LENGTH)
+dl = torch.utils.data.DataLoader(ds, batch_size=constants.BATCH_SIZE, shuffle=True, collate_fn=ds.collate_function)
 
 # Generate random sample data
 # tgt_data = torch.randint(1, tgt_vocab_size, (64, max_seq_length))  # (batch_size, seq_length)
@@ -56,26 +54,7 @@ transformer.train()
 #     optimizer.step()
 #     print(f"Epoch: {epoch+1}, Loss: {loss.item()}")
 
-def find_latest_epoch_file(path='./'):
-    epoch_files = [f for f in os.listdir(path) if re.match(r'transformer_epoch_\d+\.pt', f)]
-    if epoch_files:
-        # Extracting epoch numbers from the files and finding the max
-        latest_epoch = max([int(f.split('_')[2].split('.')[0]) for f in epoch_files])
-        return latest_epoch, f"./transformer_epoch_{latest_epoch}.pt"
-    else:
-        return 0, None
-
-# Function to load the latest epoch file if it exists
-def load_latest_checkpoint(model, path='./'):
-    latest_epoch, latest_file = find_latest_epoch_file(path)
-    if latest_file:
-        print(f"Resuming training from epoch {latest_epoch+1}")
-        model.load_state_dict(torch.load(latest_file))
-    else:
-        print("No checkpoint found, starting from beginning")
-    return latest_epoch
-
-start_epoch = load_latest_checkpoint(transformer)
+start_epoch = utilities.load_latest_checkpoint(transformer)
 
 for epoch in range(start_epoch, constants.NUM_OF_EPOCHS):
   total_loss = 0
